// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/cauHinhHoaHong.model.ts
export class CauHinhMucHoaHong {
    id: string;
    vendorMappingOptionId: string | null;
    loaiHoaHong: number | null;
    giaTriHoaHong: number | null;
    dieuKienId: string | null;
    dieuKienName: string;
    giaTriTu: number | null;
    giaTriDen: number | null;
    createdDate: string;
    parentId: string | null;

    index: number;
    indexParent: number;
}